% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEnrichment.R
\name{calculateEnrichment}
\alias{calculateEnrichment}
\title{Calculate Enrichment using EnrichR}
\usage{
calculateEnrichment(
  genes,
  enrichr_db,
  simplify = T,
  visualise = F,
  col = "#d2d2d2"
)
}
\arguments{
\item{genes}{vector of HGNC gene identifiers}

\item{enrichr_db}{Enrichr database to use, as in \url{https://maayanlab.cloud/Enrichr/#stats}}

\item{simplify}{logical; perform simplifcation calculations/filtering on GOBP and Reactome?}

\item{visualise}{logical; if T returns a ggplot object}

\item{col}{fill colour of columns in ggplot}
}
\value{
Either a dataframe  or ggplot visualisation of enriched terms
}
\description{
Calculate Enrichment using EnrichR
}
