% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleGenes.R
\docType{data}
\name{sampleGenes}
\alias{sampleGenes}
\title{A list of sample genes}
\format{
A data frame with 375 rows and 1 variables:
\describe{
  \item{gene}{HGNC gene symbol}
  ...
}
}
\source{
\url{https://maayanlab.cloud/Enrichr/#/}
}
\usage{
sampleGenes
}
\description{
A list of sample genes
}
\keyword{datasets}
