% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveEnrichmentVis.R
\name{saveEnrichmentVis}
\alias{saveEnrichmentVis}
\title{Save Enrichment Visualization}
\usage{
saveEnrichmentVis(gg, path)
}
\arguments{
\item{gg}{ggplot object}

\item{path}{image file save path}
}
\value{
Image saved to path
}
\description{
Save Enrichment Visualization
}
