% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyGO.R
\name{simplifyGO}
\alias{simplifyGO}
\title{REmove semantically similar GOBP terms}
\usage{
simplifyGO(GORes, simThresh = 0.4, adjpcol)
}
\arguments{
\item{GORes}{Output of calculateEnrichment of enrichr::enrichr functions}

\item{simThresh}{Maximum allowed semantic similarity.}

\item{adjpcol}{Column in GORes with enrichment p-value}
}
\value{
Filtered dataframe containing distinct GOBP terms
}
\description{
REmove semantically similar GOBP terms
}
