% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visEnrichment.R
\name{visEnrichment}
\alias{visEnrichment}
\title{Title}
\usage{
visEnrichment(enriched, col = "#d2d2d2")
}
\arguments{
\item{enriched}{Output dataframe of calculateEnrichment() (if visualisation = F)}

\item{col}{Fill colour}
}
\value{
ggplot object
}
\description{
Title
}
